+++ src/features/helper/types/environmentVariable.ts
@@ -1,0 +1,4 @@
+export default interface EnvironmentVariable {
+  key: string;
+  value: string;
+}